Slightly refactor terrain code to allow reuse in dumping terrain packets
